+++ lib/src/effects/color_transition_effect.dart
@@ -1,26 +1,29 @@
 import 'package:flutter/material.dart';
-import 'package:smooth_page_indicator/src/painters/expanding_dots_painter.dart';
 import 'package:smooth_page_indicator/src/painters/indicator_painter.dart';
+import 'package:smooth_page_indicator/src/painters/color_transition_painter.dart';
 
 import 'indicator_effect.dart';
 
-class ExpandingDotsEffect extends IndicatorEffect {
-  /// [expansionFactor] is multiplied by [dotWidth] to resolve
-  /// the width of the expanded dot.
-  final double expansionFactor;
+class ColorTransitionEffect extends IndicatorEffect {
+  // The active dot color
+  final Color activeDotColor;
 
-  const ExpandingDotsEffect({
-    this.expansionFactor = 3,
+  // The active dot strokeWidth
+  final double activeStrokeWidth;
+
+  const ColorTransitionEffect({
+    this.activeDotColor = Colors.indigo,
+    this.activeStrokeWidth = 1.5,
     double offset,
     double dotWidth = 16.0,
     double dotHeight = 16.0,
     double spacing = 8.0,
-    double radius = 16.0,
-    Color dotColor = Colors.indigo,
+    double radius = 16,
+    Color dotColor = Colors.grey,
     double strokeWidth = 1.0,
     PaintingStyle paintStyle = PaintingStyle.fill,
-  })  : assert(expansionFactor != null),
-        assert(expansionFactor > 1),
+  })  : assert(activeDotColor != null),
+        assert(activeStrokeWidth != null),
         super(
           dotWidth: dotWidth,
           dotHeight: dotHeight,
@@ -32,16 +35,8 @@ class ExpandingDotsEffect extends IndicatorEffect {
         );
 
   @override
-  Size calculateSize(int count) {
-    // Add the expanded dot width to our size calculation
-    return Size(
-        ((dotWidth + spacing) * (count - 1)) + (expansionFactor * dotWidth),
-        dotHeight);
-  }
-
-  @override
   IndicatorPainter buildPainter(int count, double offset, bool isRTL) {
-    return ExpandingDotsPainter(
+    return TransitionPainter(
         count: count, offset: offset, effect: this, isRTL: isRTL);
   }
 }