+++ lib/src/painters/expanding_dots_painter.dart
@@ -10,15 +10,17 @@ class ExpandingDotsPainter extends IndicatorPainter {
     @required double offset,
     @required this.effect,
     @required int count,
-  }) : super(offset, count, effect);
+    @required bool isRTL,
+  }) : super(offset, count, effect, isRTL);
 
   @override
   void paint(Canvas canvas, Size size) {
     final int current = offset.floor();
     double lastPos = -effect.spacing;
+    final dotOffset = offset - offset.toInt();
+
     for (int i = 0; i < count; i++) {
       final active = i == current;
-      final dotOffset = offset - offset.toInt();
       final bool isNext = i - 1 == current;
       final bounds =
           _calcBounds(lastPos, size.height, i, active, isNext, dotOffset);