+++ lib/src/painters/jumping_dot_painter.dart
@@ -10,7 +10,8 @@ class JumpingDotPainter extends IndicatorPainter {
     @required this.effect,
     @required int count,
     @required double offset,
-  }) : super(offset, count, effect);
+    @required bool isRTL,
+  }) : super(offset, count, effect, isRTL);
 
   @override
   void paint(Canvas canvas, Size size) {