+++ lib/src/smooth_page_indicator.dart
@@ -8,15 +8,19 @@ class SmoothPageIndicator extends AnimatedWidget {
   // a PageView controller to listen for page offset updates
   final PageController controller;
 
-  /// holds effect configuration to be used in the [IndicatorPainter]
+  /// Holds effect configuration to be used in the [IndicatorPainter]
   final IndicatorEffect effect;
 
-  /// the count of children in [PageView]
+  /// The number of children in [PageView]
   final int count;
+
+  /// If [textDirection] is [TextDirection.rtl], page offset will be reversed
+  final TextDirection textDirection;
 
   SmoothPageIndicator({
     @required this.controller,
     @required this.count,
+    this.textDirection,
     this.effect = const WormEffect(),
     Key key,
   })  : assert(controller != null),
@@ -26,12 +30,15 @@ class SmoothPageIndicator extends AnimatedWidget {
 
   @override
   Widget build(BuildContext context) {
+    // if textDirection is not provided use the nearest directionality up the widgets tree;
+    final isRTL =
+        (textDirection ?? Directionality.of(context)) == TextDirection.rtl;
     return CustomPaint(
       // different effects have different sizes
       // so we calculate size based on the provided effect
       size: effect.calculateSize(count),
       // rebuild the painter with the new offset every time it updates
-      painter: effect.buildPainter(count, controller.page ?? 0.0),
+      painter: effect.buildPainter(count, controller.page ?? 0.0, isRTL),
     );
   }
 }