+++ CHANGELOG.md
@@ -1,9 +1,21 @@
+## [0.1.2]
+
+- Add Color Transition effect
+
+### Breaking change!
+
+- Replace isRTL with textDirection
+- Directionality is now handled the flutter way instead of manually passing a bool value to isRTL property
+
 ## [0.1.1]
- * add documentation 
- * edit README file
+
+- Add documentation
+- Edit README file
 
 ## [0.1.0+1]
- * edit README file.
+
+- Edit README file.
 
 ## [0.1.0]
- * initial release.
+
+- Initial release.